+++ src/app/pages/details/instructor-resource-tab/partners/partners.html
@@ -1,7 +1,7 @@
 <template args="model">
     <div class="title-bar">
         {model.title}
-        <a class="filter-for-book" href="/partner-marketplace">See more</a>
+        <a class="filter-for-book" href="/partners">See more</a>
     </div>
     <div class="blurb-scroller">
         <div class="blurbs">
