+++ src/app/pages/details/instructor-resource-tab/partners/partners.js
@@ -25,29 +25,31 @@ class Partners extends componentType(spec) {
 
 }
 
-export default function ({dataPromise, targetEl, allies, bookAbbreviation}) {
+export default function ({dataPromise, targetEl, bookAbbreviation}) {
     dataPromise.then((pd) => {
-        function allyToPartner(ally) {
-            const partner = pd.find((pEntry) => pEntry.partner_name === ally.ally_heading) || {};
-
+        function toBlurb(partner) {
             return {
-                image: ally.ally_color_logo,
-                name: ally.ally_heading,
-                description: ally.ally_short_description,
+                image: partner.partner_logo,
+                name: partner.partner_name,
+                description: partner.short_partner_description,
                 cost: partner.affordability_cost,
                 type: partner.partner_type,
-                url: `/partner-marketplace?${ally.ally_heading}`
+                url: `/partners?${partner.partner_name}`
             };
         }
 
+        console.info('Find partners for', bookAbbreviation);
+        const forBook = pd.filter((p) => {
+            const books = p.books.split(';');
+
+            return books.includes(bookAbbreviation);
+        });
         const p = new Partners({
             el: targetEl,
             bookAbbreviation,
             model: {
                 title: pd.partner_list_label || '[Courseware partners]',
-                blurbs: allies
-                    .sort((a, b) => a.ally_heading.localeCompare(b.ally_heading))
-                    .map(allyToPartner)
+                blurbs: forBook.map(toBlurb)
             }
         });
     });
