--- src/app/pages/partner-marketplace/partner-marketplace.html
@@ -1,15 +1,0 @@
-<template args="model">
-    <section class="banner hero">
-        <div class="boxed">
-            <h1>{model.headline}</h1>
-            <div data-html="description" skip="true"></div>
-            <section class="desktop controls" skip="true"></section>
-        </div>
-        <img class="strips" src="/images/components/strips.svg" height="10" alt="" role="presentation">
-    </section>
-    <div class="padding">
-        <div class="active-filters desktop" skip="true"></div>
-        <div class="mobile controls" skip="true"></div>
-        <section class="results" skip="true"></section>
-    </div>
-</template>
