--- src/app/pages/partner-marketplace/partner-marketplace.scss
@@ -1,44 +1,0 @@
-@import 'pattern-library/core/pattern-library/headers';
-@import 'mixins/placeholder-selectors';
-
-.partner-marketplace.page {
-    .banner {
-        @extend %light-on-dark;
-
-        background-color: os-color(deep-green);
-        background-image: url('/images/partners/partner-heading-bg.png');
-        background-position: left top;
-        background-size: cover;
-        color: text-color(white);
-        padding-bottom: 8rem;
-        position: relative;
-
-        h1 {
-            margin-bottom: 3rem;
-
-            @include width-up-to($phone-max) {
-                @include set-font(h1);
-            }
-        }
-
-        .strips {
-            bottom: 0;
-        }
-    }
-
-    @include width-up-to($phone-max) {
-        .desktop {
-            display: none;
-        }
-    }
-
-    @include wider-than($phone-max) {
-        .mobile {
-            display: none;
-        }
-    }
-
-    .padding {
-        padding: 0 $normal-margin;
-    }
-}
