+++ src/app/pages/partners/active-filters/active-filters.html
