+++ src/app/pages/partners/active-filters/active-filters.scss
