+++ src/app/pages/partners/active-filters/filter-remover/filter-remover.js
