+++ src/app/pages/partners/analytics-events.js
