+++ src/app/pages/partners/controls/checkboxes-linked-to-store/checkboxes-linked-to-store.scss
