+++ src/app/pages/partners/controls/controls.html
