+++ src/app/pages/partners/controls/controls.js
