+++ src/app/pages/partners/controls/controls.scss
@@ -5,7 +5,7 @@
 $breakpt-narrow: 76.7em * $scale-factor;
 $breakpt-med: 111.9em * $scale-factor;
 
-.partner-marketplace .desktop.controls {
+.partners .desktop.controls {
     background-color: ui-color(white);
     border-bottom: thin solid ui-color(form-border);
     border-radius: 0.4rem;
