+++ src/app/pages/partners/controls/options-list/options-list.scss
