+++ src/app/pages/partners/mobile-filters/mobile-filters.scss
@@ -1,6 +1,6 @@
 @import 'pattern-library/core/pattern-library/headers';
 
-.partner-marketplace .mobile-filters {
+.partners .mobile-filters {
 
     padding: 0.5rem;
