+++ src/app/pages/partners/partner-details/carousel/carousel.html
