+++ src/app/pages/partners/partner-details/partner-details.html
