--- src/app/pages/partners/partner-viewer/blurbs.html
@@ -1,7 +1,0 @@
-<template args="model">
-    <div each="partner in model.partners" class="text" id="{partner.id+'-blurb'}">
-        <h3 data-html="partners[{$item}].title"></h3>
-        <div data-html="partners[{$item}].long_description"></div>
-        <a href="#filter" class="to-top">Return to top <i class="fa fa-chevron-up"></i></a>
-    </div>
-</template>
