--- src/app/pages/partners/partner-viewer/icons.html
@@ -1,11 +1,0 @@
-<template args="model">
-    <div each="partner in model.partners" class="logo">
-        <a class="logo-text" href="#{partner.id}-blurb">
-            <if condition="partner.ally_bw_logo">
-                <img src="{partner.ally_bw_logo}" alt="{partner.title}">
-            <else>
-                {partner.title}
-            </if>
-        </a>
-    </div>
-</template>
