+++ src/app/pages/partners/partners.scss
@@ -1,131 +1,44 @@
 @import 'pattern-library/core/pattern-library/headers';
+@import 'mixins/placeholder-selectors';
 
-.partners-page {
-    background-color: ui-color(white);
-    line-height: 3rem;
+.partners.page {
+    .banner {
+        @extend %light-on-dark;
 
-    .container {
-        align-items: left;
-
-        @include wider-than($tablet-max) {
-            align-items: center;
-        }
-    }
-
-    .hero {
-        background-color: os-color(green);
+        background-color: os-color(deep-green);
+        background-image: url('/images/partners/partner-heading-bg.png');
+        background-position: left top;
+        background-size: cover;
         color: text-color(white);
+        padding-bottom: 8rem;
+        position: relative;
 
         h1 {
-            color: inherit;
-            margin-bottom: 0.5em;
+            margin-bottom: 3rem;
+
+            @include width-up-to($phone-max) {
+                @include set-font(h1);
+            }
         }
 
-        .ally-logo {
-            display: block;
-            height: auto;
-            margin-top: 2rem;
-            width: calc(15rem + 5%);
-        }
-
-        .logo-text {
-            margin-top: 0.6em;
+        .strips {
+            bottom: 0;
         }
     }
 
-    h3 {
-        margin-bottom: 0;
-    }
-
-    .icons {
-        color: os-color(gray);
-        display: block;
-        font-weight: 500;
-        margin: 0 auto;
-        min-height: 10rem;
-        position: relative;
-        width: 100%;
-
-        a {
-            color: inherit;
-        }
-
-        .container {
-            display: grid;
-            grid-template-columns: repeat(auto-fit, minmax(22rem, 1fr));
-            justify-content: center;
-            padding-top: 6rem;
-        }
-
-        .logo {
-            align-items: center;
-            display: flex;
-            flex-basis: 22rem;
-            justify-content: center;
-            padding: 3rem 0;
-
-            &.hidden {
-                display: none;
-            }
-
-            &.spacer {
-                padding: 0;
-            }
-
-            a {
-                transition: All ease-in-out 0.3s;
-
-                &:hover,
-                &:focus {
-                    opacity: 0.7;
-                }
-            }
-
-            img {
-                max-height: 7.5rem;
-                max-width: 14rem;
-            }
+    @include width-up-to($phone-max) {
+        .desktop {
+            display: none;
         }
     }
 
-    .divider {
-        margin: 0 auto;
-
-        > hr {
-            border: 0;
-            border-bottom: solid 0.3rem ui-color(form-border);
-            width: 100%;
+    @include wider-than($phone-max) {
+        .mobile {
+            display: none;
         }
     }
 
-    .blurbs {
-        background-color: ui-color(white);
-        font-weight: 400;
-        margin-bottom: 1.5rem;
-        padding: 3rem 1.5rem;
-
-        @include width-up-to($phone-max) {
-            padding: 0 1.5rem;
-        }
-
-        .text {
-            align-items: flex-start;
-            display: flex;
-            flex-direction: column;
-            max-width: 100%;
-
-            &.hidden {
-                display: none;
-            }
-
-            p,
-            div {
-                width: 100%;
-            }
-        }
-
-        a {
-            font-weight: 500;
-        }
+    .padding {
+        padding: 0 $normal-margin;
     }
 }
