+++ src/app/pages/partners/results/result/result.html
