+++ src/app/pages/partners/results/results.html
