+++ src/app/pages/partners/results/results.js
