+++ src/app/pages/partners/results/results.scss
@@ -1,6 +1,6 @@
 @import 'pattern-library/core/pattern-library/headers';
 
-.partner-marketplace .results {
+.partners .results {
     padding: 0 0 20rem;
 
     @include wider-than($phone-max) {
