+++ src/app/pages/partners/store.js
